+++ test/FunctionsV2/PlatformSpecificHelpers.FunctionsV2.cs
@@ -177,7 +177,7 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask.Tests
 
             serviceCollection.TryAddSingleton<IApplicationLifetimeWrapper, HostLifecycleService>();
 #pragma warning disable CS0612 // Type or member is obsolete
-            serviceCollection.TryAddSingleton<IPlatformInformationService, DefaultPlatformInformationProvider>();
+            serviceCollection.TryAddSingleton<IPlatformInformation, DefaultPlatformInformation>();
 #pragma warning restore CS0612 // Type or member is obsolete
 
             return builder;