+++ release_notes.md
@@ -8,6 +8,7 @@
 * Fix fetching of large inputs for pending orchestrations on Azure Storage
 * Updated TableQuery filter condition string generation to resolve invalid character issues
 * Fixed stuck orchestration with duplicate message warning issue
+* Throw meaningful exceptions inside orchestrations when they try to call, signal, or lock a non-existing entity
 * Fixed null reference exceptions thrown in DurableClient
 
 ## Breaking Changes